requirements.txt
Django
djangorestframework
whitenoise
gunicorn
pre-commit
black
flake8